Add more tests and revert overflow error type

+++ hw4/main.c
@@ -49,12 +49,12 @@ char* convertTypeToStr(int type) {
 }
 
 // NOTE: caller needs to free returned value
-char* convertValueToStr(int64_t val, char debug) {
+char* convertValueToStr(uint64_t val, char debug) {
   int valType = getValueType(val);
   char valueStr[21];
   switch (valType) {
     case NUM_TYPE:
-      sprintf(valueStr, "%ld", val >> 1);
+      sprintf(valueStr, "%ld", ((int64_t) val) >> 1);
       break;
     case BOOL_TYPE:
       if (val == TRUE) {
